Working turn-signal state machine
Have a working implementation of the turn-signals state machine. To do it, worked on adding a timer and sampling acceleration from the mpu9250 instead of the analog accelerometer. Transitions work correctly. I also used button0 on the buckler as a substitude for ble but framework should be in place to add ble button advertisement receiving. Problems/more details in bernard_notes.txt